refac: markdown logic
